refactor: edit SEO component
